PackageGraph: Separate implementation across multiple files
